requirements.txt
fastapi==0.104.0
uvicorn==0.25.0
pydantic==1.10.13
python-dotenv==1.0.0
sqlalchemy==1.4.52
pymysql==1.1.0
google-generativeai==0.4.1
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
python-multipart==0.0.9